🐞fix: Sua loi do chay lai phan app.exec - sua thanh plt.ion()
