+++ main.py
@@ -8,6 +8,7 @@ from PyQt5.QtSql import QSqlDatabase, QSqlQueryModel, QSqlQuery
 import sqlite3 
 import matplotlib.pyplot as plt
 import pandas as pd 
+from AnalyseData import AnalyseData 
 
 def createConnection():
     con = QSqlDatabase.addDatabase("QSQLITE")
@@ -170,7 +171,7 @@ class MainWindow(QMainWindow):
         widget.setCurrentIndex(widget.currentIndex()+1)
     
     def analyse_data(self):
-        analysedata=AnalyseData()
+        analysedata=AnalyseData(widget, MainWindow)
         widget.addWidget(analysedata)
         widget.setCurrentIndex(widget.currentIndex()+1)
 
@@ -257,199 +258,6 @@ class AddData(QMainWindow):
         mainwindow = MainWindow()
         widget.addWidget(mainwindow)
         widget.setCurrentIndex(widget.currentIndex()+1)            
-
-class AnalyseData(QMainWindow):    
-    def __init__(self):
-        super(AnalyseData,self).__init__()
-        loadUi("analysedata.ui",self)
-        self.incomemonthbutton.clicked.connect(self.income_bymonth)
-        self.incometypebutton.clicked.connect(self.income_bytype)
-        self.costmonthbutton.clicked.connect(self.cost_bymonth)
-        self.costtypebutton.clicked.connect(self.cost_bytype)
-        self.incomecostbutton.clicked.connect(self.compare_incomecost)
-        self.backbutton.clicked.connect(self.back_window)
-
-    def income_bymonth(self):
-        # x-coordinates of left sides of bars  
-        left = [1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12] 
-
-        # heights of bars 
-        connection = sqlite3.connect("csdl.db")
-        sql = "SELECT SUM(income) FROM incomes GROUP BY month"
-        cursor = connection.execute(sql)
-        income_values = []
-        for row in cursor:
-            income_values.append(row[0])
-        connection.close()
-
-        # figure = plt.figure()
-        # axes = figure.add_axes([0,0,1,1])
-
-        # labels for bars 
-        tick_label = ['Jan', 'Feb', 'Mar', 'Apr', 'May', 'Jun', 'Jul', 'Aug', 'Sep', 'Oct', 'Nov', 'Dec'] 
-
-        # plotting a bar chart 
-        # axes.bar(left, income_values, tick_label = tick_label, width = 0.8, color = ['red', 'green']) 
-
-        # axes.bar(left, income_values)
-
-        # naming the x-axis 
-        # axes.xlabel('Months') 
-        # naming the y-axis 
-        # axes.ylabel('Income') 
-        # plot title 
-        # axes.title('Income over months!') 
-
-        # function to show the plot 
-        # plt.close(figure)
-        # figure.canvas.draw()
-        # plt.draw()
-
-        plt.ion()
-        fig = plt.figure(figsize=(10,8))
-        ax = fig.add_axes([0.1,0.1,0.8,0.8])
-        ax.bar(left, income_values, tick_label = tick_label, width = 0.8, color = ['red', 'green']) 
-        ax.set_title('Income over months!')
-        ax.set_xlabel('Months')
-        ax.set_ylabel('Income')
-        # raw_input('Enter to close and Continue: ')
-        # plt.close(fig)
-        # fig.canvas.draw()
-
-    def income_bytype(self):
-        connection = sqlite3.connect("csdl.db")
-        
-        # defining labels 
-        sql = "SELECT incometype FROM incomes GROUP BY incometype"
-        cursor = connection.execute(sql)
-        types = []
-        for row in cursor:
-            types.append(row[0])
-
-        # portion covered by each label
-        sql = "SELECT SUM(income) FROM incomes GROUP BY incometype"
-        cursor = connection.execute(sql)
-        slices = []
-        for row in cursor:
-            slices.append(row[0])
-        connection.close()
-
-        # plot title 
-        plt.title('Income from different activities!') 
-
-        # plotting the pie chart 
-        explode_list = []
-        start_value = 0
-        for i in range(0,len(slices)):
-            explode_list.append(start_value)
-            start_value += 0.025
-        slices.sort(reverse=True)
-        plt.pie(slices, labels = types, colors=None, startangle=0, shadow = False, explode = explode_list, radius = 1, autopct = '%1.1f%%') 
-  
-        # plotting legend 
-        plt.legend() 
-  
-        # showing the plot 
-        plt.show() 
-
-    def cost_bymonth(self):
-        # x-coordinates of left sides of bars  
-        left = [1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12] 
-  
-        # heights of bars 
-        connection = sqlite3.connect("csdl.db")
-        sql = "SELECT SUM(cost) FROM costs GROUP BY month"
-        cursor = connection.execute(sql)
-        cost_values = []
-        for row in cursor:
-            cost_values.append(row[0])
-        connection.close()
-
-        # labels for bars 
-        tick_label = ['Jan', 'Feb', 'Mar', 'Apr', 'May', 'Jun', 'Jul', 'Aug', 'Sep', 'Oct', 'Nov', 'Dec'] 
-  
-        # plotting a bar chart 
-        plt.bar(left, cost_values, tick_label = tick_label, width = 0.8, color = ['red', 'green']) 
-  
-        # naming the x-axis 
-        plt.xlabel('Months') 
-        # naming the y-axis 
-        plt.ylabel('Cost') 
-        # plot title 
-        plt.title('Cost over months!') 
-  
-        # function to show the plot 
-        plt.show()
-
-
-    def cost_bytype(self):
-        connection = sqlite3.connect("csdl.db")
-        
-        # defining labels 
-        sql = "SELECT costtype FROM costs GROUP BY costtype"
-        cursor = connection.execute(sql)
-        types = []
-        for row in cursor:
-            types.append(row[0])
-
-        # portion covered by each label
-        sql = "SELECT SUM(cost) FROM costs GROUP BY costtype"
-        cursor = connection.execute(sql)
-        slices = []
-        for row in cursor:
-            slices.append(row[0])
-
-        connection.close()
-
-        # plot title 
-        plt.title('Cost from different activities!') 
-
-        # plotting the pie chart 
-        explode_list = []
-        start_value = 0
-        for i in range(0,len(slices)):
-            explode_list.append(start_value)
-            start_value += 0.025
-        slices.sort(reverse=True)
-        plt.pie(slices, labels = types, colors=None, startangle=0, shadow = False, explode = explode_list, radius = 1, autopct = '%1.1f%%')  
-  
-        # plotting legend 
-        plt.legend() 
-  
-        # showing the plot 
-        plt.show() 
-
-
-    def compare_incomecost(self):
-        connection = sqlite3.connect("csdl.db")
-        
-        # preparing aggregated income values
-        sql = "SELECT SUM(income) FROM incomes GROUP BY month"
-        cursor = connection.execute(sql)
-        income_values = []
-        for row in cursor:
-            income_values.append(row[0])
-        
-        # preparing aggregated cost values
-        sql = "SELECT SUM(cost) FROM costs GROUP BY month"
-        cursor = connection.execute(sql)
-        cost_values = []
-        for row in cursor:
-            cost_values.append(row[0])
-        connection.close()
-        plotdata = pd.DataFrame({'Income': income_values, 'Cost': cost_values}, index = ['Jan', 'Feb', 'Mar', 'Apr', 'May', 'Jun', 'Jul', 'Aug', 'Sep', 'Oct', 'Nov', 'Dec'])    
-        plotdata.plot(kind="bar")
-        plt.title("Income versus Cost")
-        plt.xlabel("Months")
-        plt.ylabel("Income/Cost values")
-        
-        # showing the plot 
-        plt.show()
-
-    def back_window(self):
-        mainwindow = MainWindow()
-        widget.addWidget(mainwindow)
-        widget.setCurrentIndex(widget.currentIndex()+1) 
 
 class ShowIncome(QMainWindow):    
     def __init__(self):
